+++ pytest.ini
@@ -5,4 +5,4 @@ filterwarnings =
 env_files =
     test.env
 testpaths =
-    test/functional
+    tests/functional
