+++ tests/functional/adapter/utils/fixture_bool_or.py
