+++ tests/functional/adapter/utils/fixture_datediff.py
