+++ tests/functional/adapter/utils/fixture_safe_cast.py
